Fix duplicate definitions by renaming variables.
Some of the variables in l_bsc_{hl,sin,q1,q2}.c had not been made
unique with the game-specific prefix, which led to ODR
violations. Previously, some permissive compilers allowed this and
considered the tentative definitions to declare the _same_ entity, but
newer compilers rightfully reject this. The solution is to ensure that
distinct variables really are distinct.